[ana-internal] Allow minimum MTU of 1000
Allow setting the minimum allowed MTU and the initial MTU to values
larger than 1000 (lowered from 1200).
The default initial MTU and the default minimum MTU is still 1200.

This is needed for the edgetun connection that is using quic as
underlay.
If the underlay connection lowers it's MTU to around 1200 the
edgetun connection will not be able to send or receive packets.